Refactor remote command
Fix bad command line - "--timeout null"
Fix bad command line - lunching from Graphite

Use ES6 imports
Added missing dependency in package.json - shelljs

Fixes http://teampulse.telerik.com/view#item/298968
