+++ lib/options.ts
@@ -2,8 +2,8 @@
 "use strict";
 
 import commonOptionsLibPath = require("./common/options");
-import osenv = require("osenv");
-import path = require("path");
+import * as osenv from "osenv";
+import * as path  from "path";
 
 let OptionType = commonOptionsLibPath.OptionType;
 
@@ -44,7 +44,7 @@ export class Options extends commonOptionsLibPath.OptionsBase {
 		},
 		path.join($hostInfo.isWindows ? process.env.LocalAppData : path.join(osenv.home(), ".local/share"), "Telerik", "BlackDragon", ".appbuilder-cli"),
 			$errors, $staticConfig);
-		
+
 		let that = <any>this;
 		that.screenBuilderCacheDir = path.join((($hostInfo.isWindows && this.defaultProfileDir === that.profileDir) ? path.join(process.env.LocalAppData, "Telerik"): that.profileDir), "sb");
 	}
